review update

+++ backend/src/main/kotlin/com/group7/artshare/service/EventService.kt
@@ -20,69 +20,64 @@ class EventService(
     private val imageRepository: ImageRepository
 ) {
     fun createPhysicalExhibition(
-        physicalExhibitionRequest: PhysicalExhibitionRequest,
-        user: RegisteredUser
+        physicalExhibitionRequest: PhysicalExhibitionRequest, user: RegisteredUser
     ): PhysicalExhibition {
         val newPhysicalExhibition = PhysicalExhibition()
         if (user is Artist) {
             newPhysicalExhibition.creator = user
             user.hostedEvents.add(newPhysicalExhibition)
-        } else
-            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Registered users cannot create physical exhibitions")
+        } else throw ResponseStatusException(
+            HttpStatus.BAD_REQUEST,
+            "Registered users cannot create physical exhibitions"
+        )
         newPhysicalExhibition.location = physicalExhibitionRequest.location
         newPhysicalExhibition.rules = physicalExhibitionRequest.rules
-        if (physicalExhibitionRequest.eventInfo?.posterId?.let { imageRepository.existsById(it) } == false)
-            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "There is no image in the database with this id")
+        if (physicalExhibitionRequest.eventInfo?.posterId?.let { imageRepository.existsById(it) } == false) throw ResponseStatusException(
+            HttpStatus.BAD_REQUEST,
+            "There is no image in the database with this id"
+        )
         newPhysicalExhibition.eventInfo = physicalExhibitionRequest.eventInfo
         physicalExhibitionRepository.save(newPhysicalExhibition)
-        //return PhysicalExhibition()
         return newPhysicalExhibition
     }
 
     fun createOnlineGallery(onlineGalleryRequest: OnlineGalleryRequest, user: RegisteredUser): OnlineGallery {
         val newOnlineGallery = OnlineGallery()
+        if (onlineGalleryRequest.eventInfo?.posterId?.let { imageRepository.existsById(it) } == false) throw ResponseStatusException(
+            HttpStatus.BAD_REQUEST,
+            "There is no image in the database with this id"
+        )
+        newOnlineGallery.eventInfo = onlineGalleryRequest.eventInfo
+        onlineGalleryRequest.artItemIds?.forEach {
+            val artItem = artItemRepository.findByIdOrNull(it) ?: throw ResponseStatusException(
+                HttpStatus.BAD_REQUEST, "Art item list includes an id with no corresponding art item"
+            )
+            newOnlineGallery.artItems.add(artItem)
+        }
         if (user is Artist) {
             newOnlineGallery.creator = user
             user.hostedEvents.add(newOnlineGallery)
-        } else
-            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Registered users cannot create online galleries")
-        if (onlineGalleryRequest.eventInfo?.posterId?.let { imageRepository.existsById(it) } == false)
-            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "There is no image in the database with this id")
-        newOnlineGallery.eventInfo = onlineGalleryRequest.eventInfo
-        onlineGalleryRequest.artItemIds?.forEach {
-            val artItem = artItemRepository.findByIdOrNull(it)
-                ?: throw ResponseStatusException(
-                    HttpStatus.BAD_REQUEST,
-                    "Art item list includes an id with no corresponding art item"
-                )
-            newOnlineGallery.artItems.add(artItem)
-        }
+        } else throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Regular users cannot create online galleries")
         onlineGalleryRepository.save(newOnlineGallery)
         return newOnlineGallery
     }
 
     fun deleteEvent(
-        id: Long,
-        user: RegisteredUser
+        id: Long, user: RegisteredUser
     ) {
         if (!((physicalExhibitionRepository.existsById(id)) || onlineGalleryRepository.existsById(id))) throw ResponseStatusException(
-            HttpStatus.BAD_REQUEST,
-            "There is no event in the database with corresponding id"
+            HttpStatus.BAD_REQUEST, "There is no event in the database with corresponding id"
         )
         if (user is Artist) {
-            var event: Event? = user.hostedEvents.firstOrNull { it.id == id }
-                ?: throw ResponseStatusException(HttpStatus.UNAUTHORIZED, "Event does not belong to this user")
-            if (event != null) {
-                user.hostedEvents.remove(event)
-                if (event is PhysicalExhibition)
-                    physicalExhibitionRepository.delete(event)
-                if (event is OnlineGallery)
-                    onlineGalleryRepository.delete(event)
-            }
+            var event: Event = user.hostedEvents.firstOrNull { it.id == id } ?: throw ResponseStatusException(
+                HttpStatus.UNAUTHORIZED,
+                "Event does not belong to this user"
+            )
+            user.hostedEvents.remove(event)
+            if (event is PhysicalExhibition) physicalExhibitionRepository.delete(event)
+            if (event is OnlineGallery) onlineGalleryRepository.delete(event)
         } else {
             throw ResponseStatusException(HttpStatus.UNAUTHORIZED, "Regular Users cannot delete events")
         }
     }
-
-
 }
